Addressing more comments
