+++ src/drtsans/tof/eqsans/inelastic_correction.py
@@ -1,5 +1,5 @@
 """
-This module provides the functionality to correct I(Q, wavelength) accounting for wavelength-dependent incoherent
+This module provides the functionality to correct I(Q, wavelength) accounting for wavelength-dependent
 inelastic scattering, for both 1D and 2D data.
 """
 
@@ -12,9 +12,10 @@ from drtsans.tof.eqsans.elastic_correction import (
     determine_reference_wavelength_intensity_mesh,
     reshape_intensity_domain_meshgrid,
 )
-from drtsans.dataobjects import getDataType, DataType, save_i1d, IQazimuthal
+from drtsans.dataobjects import getDataType, DataType, save_i1d, IQazimuthal, IQmod
 from collections import namedtuple
 from mantid.kernel import logger
+from typing import Tuple, List, Optional
 import numpy as np
 import os
 
@@ -28,6 +29,7 @@ __all__ = [
     "CorrectedI1D",
     "CorrectedIQ2D",
     "CorrectionFactors",
+    "inelastic_correction",
 ]
 
 # Output of corrected 1D case
@@ -814,3 +816,161 @@ def correct_intensity_error(
         corrected_errors[:, i_wl] = term1
 
     return corrected_intensities, np.sqrt(corrected_errors)
+
+
+def inelastic_correction(
+    iq2d_unbinned: IQazimuthal,
+    iq1d_unbinned: IQmod,
+    num_x_bins: int,
+    num_y_bins: int,
+    num_q1d_bins: int,
+    num_q1d_bins_per_decade: int,
+    decade_on_center: bool,
+    bin1d_type: str,
+    log_binning: bool,
+    user_qmin: Optional[float],
+    user_qmax: Optional[float],
+    annular_bin: float,
+    wedges: List[Tuple[int, int]],
+    symmetric_wedges: bool,
+    weighted_errors: bool,
+    select_min_incoherence: bool,
+    intensity_weighted: bool,
+    incoh_qmin: Optional[float],
+    incoh_qmax: Optional[float],
+    incoh_factor: Optional[float],
+    output_dir: str,
+    output_filename: str,
+    raw_name: str,
+) -> Tuple[IQazimuthal, IQmod]:
+    """Apply inelastic/incoherence correction to unbinned I(Q) data.
+
+    This function encapsulates all inelastic correction logic:
+    1. Determines Q ranges from data if not provided by user
+    2. Temporarily bins data in scalar mode to calculate b(λ) factors
+    3. Saves b(λ) to file
+    4. Applies b(λ) correction to unbinned sample data
+    5. Returns corrected unbinned data
+
+    Parameters
+    ----------
+    iq2d_unbinned : IQazimuthal
+        Unbinned 2D I(Qx, Qy) sample data
+    iq1d_unbinned : IQmod
+        Unbinned 1D I(Q) sample data
+    num_x_bins : int
+        Number of Qx bins for temporary binning
+    num_y_bins : int
+        Number of Qy bins for temporary binning
+    num_q1d_bins : int
+        Number of Q bins for temporary 1D binning
+    num_q1d_bins_per_decade : int
+        Number of bins per decade for logarithmic binning
+    decade_on_center : bool
+        Whether decade boundaries are on bin centers
+    bin1d_type : str
+        Type of 1D binning ('scalar', 'wedge', 'annular')
+    log_binning : bool
+        Whether to use logarithmic binning
+    user_qmin : float, optional
+        Minimum Q value for binning (if None, determined from data)
+    user_qmax : float, optional
+        Maximum Q value for binning (if None, determined from data)
+    annular_bin : float
+        Width of annular bin in degrees
+    wedges : list
+        List of (angle_min, angle_max) tuples for wedges
+    symmetric_wedges : bool
+        Whether to add symmetric wedges
+    weighted_errors : bool
+        Whether to use error-weighted binning
+    select_min_incoherence : bool
+        Flag to determine correction B by minimum incoherence
+    intensity_weighted : bool
+        Whether to use intensity-weighted B factor calculation
+    incoh_qmin : float, optional
+        Minimum Q for incoherence correction calculation
+    incoh_qmax : float, optional
+        Maximum Q for incoherence correction calculation
+    incoh_factor : float, optional
+        Factor for automatic Q range determination
+    output_dir : str
+        Output directory for correction files
+    output_filename : str
+        Base filename for output
+    raw_name : str
+        Prefix for correction file names
+
+    Returns
+    -------
+    tuple
+        (corrected_iq2d_unbinned, corrected_iq1d_unbinned)
+    """
+    # Import bin_all here to avoid circular import
+    from drtsans.iq import bin_all
+    from drtsans.tof.eqsans.correction_api import save_b_factor
+
+    logger.notice("Applying inelastic/incoherent correction")
+
+    # Determine Q ranges from data if not provided by user
+    qmin = user_qmin if user_qmin is not None else iq1d_unbinned.mod_q.min()
+    qmax = user_qmax if user_qmax is not None else iq1d_unbinned.mod_q.max()
+    qxrange = (np.min(iq2d_unbinned.qx), np.max(iq2d_unbinned.qx))
+    qyrange = (np.min(iq2d_unbinned.qy), np.max(iq2d_unbinned.qy))
+
+    # Temporarily bin data in scalar mode to calculate b(λ) factors
+    # These binned results are ONLY used for factor calculation, then discarded
+    iq2d_temp_binned, iq1d_temp_binned = bin_all(
+        iq2d_unbinned,
+        iq1d_unbinned,
+        num_x_bins,
+        num_y_bins,
+        n1dbins=num_q1d_bins,
+        n1dbins_per_decade=num_q1d_bins_per_decade,
+        decade_on_center=decade_on_center,
+        bin1d_type="scalar" if bin1d_type == "wedge" else bin1d_type,
+        log_scale=log_binning,
+        qmin=qmin,
+        qmax=qmax,
+        qxrange=qxrange,
+        qyrange=qyrange,
+        annular_angle_bin=annular_bin,
+        wedges=wedges,
+        symmetric_wedges=symmetric_wedges,
+        error_weighted=weighted_errors,
+        n_wavelength_bin=None,
+    )
+
+    if len(iq1d_temp_binned) != 1:
+        raise NotImplementedError("Expected exactly one IQmod from temporary binning")
+
+    # Calculate b(λ) correction factors from scalar-binned I(Q, λ)
+    logger.notice("Calculating inelastic/incoherent correction factors from scalar-binned I(Q, lambda)")
+    correction_factors = calculate_incoherence_correction_factors(
+        iq1d_temp_binned[0],
+        select_min_incoherence,
+        intensity_weighted,
+        incoh_qmin,
+        incoh_qmax,
+        incoh_factor,
+    )
+
+    # Save b(λ) to file
+    inelastic_output_dir = os.path.join(output_dir, "info", "inelastic_incoh", output_filename)
+    os.makedirs(inelastic_output_dir, exist_ok=True)
+
+    WavelengthContainer = namedtuple("WavelengthContainer", ["wavelength"])
+    wl_container = WavelengthContainer(wavelength=correction_factors.wavelength)
+
+    save_b_factor(
+        CorrectedI1D(wl_container, correction_factors.b_factor, correction_factors.b_error),
+        os.path.join(inelastic_output_dir, f"{output_filename}_inelastic_b1d_{raw_name}.dat"),
+    )
+
+    # Apply b(λ) to unbinned sample data
+    logger.notice("Applying inelastic/incoherent correction to unbinned data")
+    iq2d_corrected, iq1d_corrected = apply_incoherence_correction_to_unbinned_data(
+        iq2d_unbinned, iq1d_unbinned, correction_factors
+    )
+
+    return iq2d_corrected, iq1d_corrected
